refactor(advisor): split review responsibilities
