+++ test/pr-review-advisor-ledger-tools.test.ts
@@ -5,22 +5,13 @@ import type { ToolDefinition } from "@earendil-works/pi-coding-agent";
 import { Check } from "typebox/value";
 import { describe, expect, it } from "vitest";
 import { buildRiskPlan } from "../tools/advisors/risk-plan.mts";
-import {
-  canonicalRetryFallback,
-  normalizeReviewResult,
-  partialLedgerFailureResult,
-  recordSynthesisValidationFailureOnDraft,
-  reviewLedgerConsistencyIssues,
-  terminologyReviewConsistencyIssues,
-  withCanonicalReviewLedgerFindings,
-} from "../tools/pr-review-advisor/analyze.mts";
+import { normalizeReviewResult } from "../tools/pr-review-advisor/analyze.mts";
 import {
   createReviewFindingLedger,
   createReviewLedgerToolController,
   REVIEW_LEDGER_READ_TOOL,
   REVIEW_LEDGER_UPDATE_TOOL,
 } from "../tools/pr-review-advisor/review-ledger.mts";
-import { createTerminologyLedger } from "../tools/pr-review-advisor/terminology.mts";
 
 type CallableTool = ToolDefinition & {
   execute(
@@ -136,9 +127,7 @@ describe("PR review ledger tools", () => {
     ).toThrow("sourceOfTruthReview[1] must include findingId");
   });
 
-  it("keeps source-of-truth prose from creating findings outside the ledger", () => {
-    const ledger = createReviewFindingLedger();
-    ledger.applyBatch([{ operation: "add", finding: finding() }], "correctness-state");
+  it("keeps source-of-truth prose from creating findings", () => {
     const result = normalizeReviewResult(
       {
         findings: [{ ...finding(), evidence: finding().evidence.join("\n") }],
@@ -160,176 +149,9 @@ describe("PR review ledger tools", () => {
     );
 
     expect(result.findings).toHaveLength(1);
-    expect(reviewLedgerConsistencyIssues(result, ledger.snapshot())).toEqual([]);
-  });
-
-  it("rejects unresolved source-of-truth review without an open ledger finding", () => {
-    const result = normalizeReviewResult(
-      {
-        findings: [],
-        sourceOfTruthReview: [
-          {
-            surface: "best-effort cleanup",
-            status: "missing",
-            findingId: null,
-            invalidState: "A failed resource may remain allocated.",
-            sourceBoundary: "Resource creation lifecycle.",
-            whyNotSourceFix: "Not established.",
-            regressionTest: "Missing.",
-            removalCondition: "Unknown.",
-            evidence: "The cleanup suppresses deletion failures.",
-          },
-        ],
-      },
-      reviewMetadata(),
-    );
-
-    const snapshot = createReviewFindingLedger().snapshot();
-    expect(reviewLedgerConsistencyIssues(result, snapshot)).toEqual([
-      "sourceOfTruthReview[1] best-effort cleanup must reference an open ledger finding",
+    expect(result.sourceOfTruthReview).toMatchObject([
+      { surface: "best-effort refusal cleanup", findingId: "F-001" },
     ]);
-    expect(canonicalRetryFallback(result, snapshot)).toBeNull();
-  });
-
-  it("rejects draft finding drift before applying canonical ledger findings", () => {
-    const ledger = createReviewFindingLedger();
-    ledger.applyBatch([{ operation: "add", finding: finding() }], "correctness-state");
-    const draft = normalizeReviewResult({ findings: [] }, reviewMetadata());
-
-    expect(canonicalRetryFallback(draft, ledger.snapshot())).toBeNull();
-  });
-
-  it("accepts equivalent terminology receipts with reordered object keys", () => {
-    const metadata = reviewMetadata();
-    const findingSnapshot = createReviewFindingLedger().snapshot();
-    const terminologySnapshot = createTerminologyLedger(metadata.headSha).snapshot();
-    const normalized = normalizeReviewResult(
-      { terminologyReview: terminologySnapshot.review },
-      metadata,
-    );
-    const reordered = {
-      ...normalized,
-      terminologyReview: {
-        noChangesReason: terminologySnapshot.review.noChangesReason,
-        decisions: terminologySnapshot.review.decisions,
-        status: terminologySnapshot.review.status,
-      },
-    };
-
-    expect(terminologyReviewConsistencyIssues(reordered, terminologySnapshot)).toEqual([]);
-    expect(canonicalRetryFallback(reordered, findingSnapshot, terminologySnapshot)).not.toBeNull();
-  });
-
-  it("preserves canonical findings when a later advisor stage fails", () => {
-    const ledger = createReviewFindingLedger();
-    ledger.applyBatch([{ operation: "add", finding: finding() }], "correctness-state");
-
-    const result = partialLedgerFailureResult(
-      reviewMetadata(),
-      "tests-regressions omitted its ledger commit",
-      ledger.snapshot(),
-    );
-
-    expect(result).toMatchObject({
-      summary: { confidence: "low", recommendation: "info_only" },
-      findings: [{ title: finding().title }],
-      reviewCompleteness: { requiresHumanReview: true },
-    });
-    expect(result?.findings[0]?.title).not.toBe("PR review advisor unavailable");
-    expect(result?.reviewCompleteness.limitations[0]).toContain(
-      "stopped before completing all review stages",
-    );
-  });
-
-  it("maps a completed empty canonical ledger to merge_as_is without waiving human review", () => {
-    const result = normalizeReviewResult(
-      {
-        summary: {
-          recommendation: "info_only",
-          confidence: "high",
-          oneLine: "No actionable findings remain.",
-        },
-        findings: [],
-        reviewCompleteness: {
-          limitations: [],
-          requiresHumanReview: false,
-        },
-      },
-      reviewMetadata(),
-    );
-
-    const canonical = withCanonicalReviewLedgerFindings(
-      result,
-      createReviewFindingLedger().snapshot(),
-    );
-
-    expect(canonical.summary).toMatchObject({
-      confidence: "high",
-      recommendation: "merge_as_is",
-      oneLine: "No actionable findings remain in the canonical review ledger.",
-    });
-    expect(canonical.reviewCompleteness.requiresHumanReview).toBe(true);
-  });
-
-  it("keeps a low-confidence empty canonical ledger fail-closed as info_only (#7521)", () => {
-    const result = normalizeReviewResult(
-      {
-        summary: {
-          recommendation: "merge_as_is",
-          confidence: "low",
-          oneLine: "No actionable findings remain.",
-        },
-        findings: [],
-        reviewCompleteness: {
-          limitations: ["Review confidence remained low."],
-          requiresHumanReview: true,
-        },
-      },
-      reviewMetadata(),
-    );
-
-    const canonical = withCanonicalReviewLedgerFindings(
-      result,
-      createReviewFindingLedger().snapshot(),
-    );
-
-    expect(canonical.summary).toMatchObject({
-      confidence: "low",
-      recommendation: "info_only",
-    });
-    expect(canonical.reviewCompleteness.requiresHumanReview).toBe(true);
-  });
-
-  it("marks a canonical draft incomplete when synthesis validation fails (#7521)", () => {
-    const result = normalizeReviewResult(
-      {
-        summary: {
-          recommendation: "merge_as_is",
-          confidence: "high",
-          oneLine: "No actionable findings remain.",
-        },
-        findings: [],
-        reviewCompleteness: {
-          limitations: [],
-          requiresHumanReview: false,
-        },
-      },
-      reviewMetadata(),
-    );
-    const canonical = canonicalRetryFallback(result, createReviewFindingLedger().snapshot());
-
-    expect(canonical).not.toBeNull();
-    const fallback = recordSynthesisValidationFailureOnDraft(canonical!, "validation turn failed");
-
-    expect(fallback.summary).toMatchObject({
-      confidence: "low",
-      recommendation: "info_only",
-      oneLine: "Same-session synthesis validation failed; the advisor result is incomplete.",
-    });
-    expect(fallback.reviewCompleteness.requiresHumanReview).toBe(true);
-    expect(fallback.reviewCompleteness.limitations).toContain(
-      "Same-session synthesis validation failed; using canonical draft: validation turn failed",
-    );
   });
 
   it("binds mutations to the runner stage and exposes the canonical snapshot (#6446)", async () => {
@@ -777,44 +599,6 @@ describe("PR review ledger tools", () => {
     expect(ledger.snapshot().findings).toMatchObject([
       { id: "F-001", title: candidate.title, status: "open" },
     ]);
-  });
-
-  it("detects synthesis drift and publishes the ledger's canonical finding (#6446)", () => {
-    const ledger = createReviewFindingLedger();
-    ledger.applyBatch([{ operation: "add", finding: finding() }], "correctness-state");
-    const drifted = {
-      summary: {
-        recommendation: "merge_as_is",
-        confidence: "high",
-        oneLine: "No findings.",
-      },
-      findings: [
-        {
-          severity: "suggestion",
-          category: "correctness",
-          file: "src/lib/runner.ts",
-          line: 42,
-          title: "Refusal status is masked",
-          description: "The refusal path returns success.",
-          impact: "Automation can treat a rejected action as successful.",
-          recommendation: "Propagate the refusal status.",
-          verificationHint: "Read the refusal return at src/lib/runner.ts:42.",
-          missingRegressionTest: "Assert that refusal returns a nonzero status.",
-          evidence: "src/lib/runner.ts:42 returns zero on refusal",
-        },
-      ],
-    } as unknown as Parameters<typeof reviewLedgerConsistencyIssues>[0];
-
-    expect(reviewLedgerConsistencyIssues(drifted, ledger.snapshot())).toEqual([
-      "final findings[1] diverges from canonical ledger finding F-001",
-    ]);
-    expect(
-      withCanonicalReviewLedgerFindings(drifted, ledger.snapshot()).findings[0]?.severity,
-    ).toBe("warning");
-    expect(withCanonicalReviewLedgerFindings(drifted, ledger.snapshot()).summary).toMatchObject({
-      recommendation: "merge_after_fixes",
-      topItem: "Refusal status is masked",
-    });
   });
 
   it("requires a reason and new evidence to change a conclusion (#6446)", () => {
